+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Made Layout::current_layer public for getting current active layer.
 * Added a procedural macro for defining layouts (`keyberon::layout::layout`)
 * Corrected HID report descriptor
+* Add max_packet_size() to HidDevice to allow differing report sizes
 
 Breaking changes:
 * Update to generic_array 0.14, which is exposed in matrix. The update